moved consumergroupmetadata to consumer package
